Refactor converters - part 2.

+++ org.jhotdraw8.css/src/main/java/org.jhotdraw8.css/org/jhotdraw8/css/parser/CssToken.java
@@ -345,7 +345,7 @@ public class CssToken {
     }
 
     private String fromNUMBER() {
-        return NUMBER_CONVERTER.toString(numericValue);
+        return NUMBER_CONVERTER.toString(numericValue.doubleValue());
     }
 
     private @NonNull String fromPERCENTAGE() {
